requirements.txt
python-dotenv~=0.13.0
Flask~=1.1.2
marshmallow~=3.6.1
requests~=2.24.0